ugly but working version of BorrowList and unusable (but non-breaking) version of RequestList
